Added 2 new services:
1) XMLService is responsible for generating nodes and getting information about them from xml file
2) FileService is responsible for creating / deleting / updating files.

Working with the files themselves is not done through Entity Core, since there is no filestream support for this technology. So you have to use part of ADO.NET.

Updates:
AccountService has undergone a change, registration has been removed from this class, now it is described inside the AccountController.

Document - Transaction_Context has been added, this field is intended for the service to work with files.

SqlFileContext is closely related to the added services.

User has got new fields.

!!! You should run the migration ProgmanDBUsersUpdateMigration